Fix bug from QA (part 3)

+++ database/migrations/2024_08_26_063031_create_transaction_details_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreign('transactionID')->references('id')->on('transaksis')->onUpdate('cascade')->onDelete('cascade');
             $table->unsignedBigInteger('productID');
             $table->foreign('productID')->references('id')->on('products')->onUpdate('cascade')->onDelete('cascade');
-            $table->integer('productQuantity');
+            $table->bigInteger('productQuantity');
         });
     }
